Изменение типа для первичного ключа на Integer

+++ evm-stat-service/stats-dto/src/main/java/ru/practicum/evm/dto/EndpointHitDto.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class EndpointHitDto {
 
-    private Long id; //Идентификатор записи
+    private Integer id; //Идентификатор записи
 
     @NotNull
     private String app; //Идентификатор сервиса для которого записывается информация
